Adding ccf-eui icons
The eui tool tries to load the icon urls relative to the base url.
Changing the base url in the head for the EUI react component causes the
SenNet icon and login/logout to break. The best solution is to add the icons
to the public folder and serve them with the portal.